+++ app/Transformers/InventoryValuationRowTransformer.php
@@ -14,7 +14,7 @@ final class InventoryValuationRowTransformer
             'part_id' => $row['part_id'] ?? null,
             'code' => $row['code'] ?? null,
             'name' => $row['name'] ?? null,
-            'quantity' => (int) ($row['quantity'] ?? 0),
+            'quantity' => (float) ($row['quantity'] ?? 0),
             'value_cost' => (float) ($row['value_cost'] ?? 0),
             'value_sell' => (float) ($row['value_sell'] ?? 0),
         ];
