+++ app/Transformers/LowStockAlertTransformer.php
@@ -12,7 +12,7 @@ final class LowStockAlertTransformer
         return [
             'part_id' => $r->part_id ?? null,
             'branch_id' => $r->branch_id ?? null,
-            'quantity' => isset($r->quantity) ? (int) $r->quantity : null,
+            'quantity' => isset($r->quantity) ? (float) $r->quantity : null,
             'min_stock' => isset($r->min_stock) ? (int) $r->min_stock : null,
             'part_name' => $r->name ?? null,
         ];
